+++ tests/test_helper.py
@@ -97,6 +97,7 @@ class TestHelper(object):
     adyen_merchant_account_id = "adyen_ma"
     hiper_brl_merchant_account_id = "hiper_brl"
     card_processor_brl_merchant_account_id = "card_processor_brl"
+    aib_swe_ma_merchant_account_id = "aib_swe_ma"
 
     add_on_discount_plan = {
          "description": "Plan for integration tests -- with add-ons and discounts",